Make wgrib2 actually return just one record
Now wgrib2 will always return the 1st record processed after matching
the variable, level, etc. Before, using the syntax `match "^1"`, it
would only match the record number 1, so if the first record found had
an id of 2, it was then discarded.

The `-end` option stops wgrib2 after it processes the first record
matched, which is what we want to avoid duplicate records.