Accept the full ensemble array in make_poe
Now the full ensemble array (all members and grid points) can be setn to
`make_poe()`. Ran it and it now takes less time to complete. However,
when `num_xvals` is set to 1000, there's a MemoryError. Settings it to
100 makes that go away. The resulting plot is very close to the one from
the master branch, though there are slight differences, assumingly due
to the change in `num_xvals`. Further diagnosis needs to happen to
determine how to not get a MemoryError when setting `num_xvals` to 1000.