Switch to using YAML instead of INI
Use PyYaml to read settings from a YAML config file instead of an INI
config file. YAML is easier to read/type, and the result of reading the
file is exactly the same as with the INI file (Python objects can be
represented in both [strings, lists, etc.], results in a Python
dictionary).