Use NumPy arange to create lats/lons
Before the Grid class used Python's built-in `range()` function to
create a list of lats and lons. This broke when specifying the grid name
`1/6th_deg_global` because the resolution is not an integer, and the
`range()` function only works with integers. We now use NumPy's
`arange()` and `tolist()` functions to generate the list and convert
back to a Python list.