Add unit conversion module
A lot of stuff is in this commit, sorry!

When I realized I needed a unit conversion module, I took the opportunity to learn how to write Python unit tests (using the Nose framework). This is the reason for tests/test_UnitConverter.py.

I also added a makefile to make nosetests-running easier, and for creating API documentation (including the pdoc templates under lib/api-doc-templates).