Test if input is masked by trying to access mask
Before, just calling getmask() on a MaskedArray would return False if
the entire mask was false, but I wanted that case to be true because
there's still a mask, it's just False everywhere.

This method will attempt to access the mask attribute, and fail with an
AttributeError if it can't...which is what I want.