Add debug arg to all functions
Whether logging level is set to DEBUG should be a function parameter, it
should not inherit from the root logger. This is because users may use
this function, not knowing that debug messages will be printed to the
screen, and no one wants to see thousands of logging messages they
weren't expecting!