Change to storing loaded data in Dataset obj
A Dataset class now exists that stores loaded data. By default all the
attributes are set to None, and only the data that gets loaded is stored
in the appropriate attribute. This was done so that eventually some QC
attributes can be saved so the end user can use QC information to make
decisions about the loaded data.